+++ lib/KSW.ATE01.Project.Base/Models/TestPlans/LimitsModel.cs
@@ -20,27 +20,28 @@ namespace KSW.ATE01.Project.Base.Models.TestPlans
 
         /// <summary>
         /// 测试项名
-        /// </summary>
+        /// </summary>       
         public string TestItemName { get; set; }
 
         /// <summary>
         /// 电压限制名称
         /// </summary>
         public string LimitName { get; set; }
+
         /// <summary>
         /// 测试编号
         /// </summary>
-        public int TestNumber { get; set; }
+        public uint TestNumber { get; set; }
 
         /// <summary>
         /// 电压下限
         /// </summary>
-        public decimal LowLimit { get; set; }
+        public double LowLimit { get; set; }
 
         /// <summary>
         /// 电压上限
         /// </summary>
-        public decimal HighLimit { get; set; }
+        public double HighLimit { get; set; }
 
         /// <summary>
         /// 单位
@@ -50,22 +51,22 @@ namespace KSW.ATE01.Project.Base.Models.TestPlans
         /// <summary>
         /// 软件Bin号
         /// </summary>
-        public int FailSoftwareBin { get; set; }
+        public uint FailSoftwareBin { get; set; }
 
         /// <summary>
         /// 软件Bin号
         /// </summary>
-        public int PassSoftwareBin { get; set; }
+        public uint PassSoftwareBin { get; set; }
 
         /// <summary>
         /// 硬件Bin号
         /// </summary>
-        public int FailHardwareBin { get; set; }
+        public uint FailHardwareBin { get; set; }
 
         /// <summary>
         /// 硬件Bin号
         /// </summary>
-        public int PassHardwareBin { get; set; }
+        public uint PassHardwareBin { get; set; }
 
         /// <summary>
         /// 测试结果
